make square_feet integer

+++ storeshare_api/db/migrate/20221102030051_create_listings.rb
@@ -10,7 +10,7 @@ class CreateListings < ActiveRecord::Migration[7.0]
       t.string :city
       t.string :state
       t.string :zip_code
-      t.decimal :square_feet
+      t.integer :square_feet
 
       t.timestamps
     end
